Quick cleanup and rearrange to make it easy to install and run
