+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/avgjoesecurity/CisternAttack
+
+go 1.12
+
+require golang.org/x/crypto v0.0.0-20190829043050-9756ffdc2472
