+++ ssh/ssh_server.go
@@ -3,14 +3,14 @@
 package main
 
 import (
-    //"io/ioutil"
-    "log"
-    "net"
-    "os"
-    "fmt"
+	//"io/ioutil"
+	"fmt"
+	"log"
+	"net"
+	"os"
 
-    "golang.org/x/crypto/ssh"
-    "encoding/base64"
+	"encoding/base64"
+	"golang.org/x/crypto/ssh"
 )
 
 var HOST_PRIVATE_KEY = []byte(`-----BEGIN RSA PRIVATE KEY-----
@@ -42,48 +42,49 @@ OtJo3ICseP7/m1ijLZKkUm0GbpKyyc50COyTEzdrgwhcfKX9ZfhD7FLo11V2EChe
 -----END RSA PRIVATE KEY-----`)
 
 func handle(conn net.Conn, config *ssh.ServerConfig) {
-        // From a standard TCP connection to an encrypted SSH connection
-        sshConn, _, _, err := ssh.NewServerConn(conn, config)
-        if err != nil {
-            log.Printf("Failed to handshake (%s)", err)
-            return
-        }
+	// From a standard TCP connection to an encrypted SSH connection
+	sshConn, _, _, err := ssh.NewServerConn(conn, config)
+	if err != nil {
+		log.Printf("Failed to handshake (%s)", err)
+		return
+	}
 
-        log.Println("Connection from", sshConn.RemoteAddr())
-        sshConn.Close()
+	log.Println("Connection from", sshConn.RemoteAddr())
+	sshConn.Close()
 }
 
 func main() {
-    hostPrivateKeySigner, err := ssh.ParsePrivateKey(HOST_PRIVATE_KEY)
-    if err != nil {
-        panic(err)
-    }
+	hostPrivateKeySigner, err := ssh.ParsePrivateKey(HOST_PRIVATE_KEY)
+	if err != nil {
+		panic(err)
+	}
 
-    config := &ssh.ServerConfig{
-        //Define a function to run when a client attempts a password login
-        PasswordCallback: func(c ssh.ConnMetadata, pass []byte) (*ssh.Permissions, error) {
-            // Should use constant-time compare (or better, salt+hash) in a production setting.
-            fmt.Println("User: "+string(c.User())+" Password: "+string(pass))
-            return nil, fmt.Errorf("password rejected for %q", c.User())
-        },
-        PublicKeyCallback: func(c ssh.ConnMetadata, key ssh.PublicKey) (*ssh.Permissions, error) {
-            fmt.Println("User: "+string(c.User())+" Public Key Type: "+string(key.Type())+" Public Key: "+(base64.StdEncoding.EncodeToString(key.Marshal())))
-            return nil, fmt.Errorf("key rejected for %q", c.User())
-        },
-    }
-    config.AddHostKey(hostPrivateKeySigner)
+	config := &ssh.ServerConfig{
+		//Define a function to run when a client attempts a password login
+		PasswordCallback: func(c ssh.ConnMetadata, pass []byte) (*ssh.Permissions, error) {
+			// Should use constant-time compare (or better, salt+hash) in a production setting.
+			fmt.Println("User: " + string(c.User()) + " Password: " + string(pass))
+			return nil, fmt.Errorf("password rejected for %q", c.User())
+		},
+		PublicKeyCallback: func(c ssh.ConnMetadata, key ssh.PublicKey) (*ssh.Permissions, error) {
+			fmt.Println("User: " + string(c.User()) + " Public Key Type: " + string(key.Type()) + " Public Key: " + (base64.StdEncoding.EncodeToString(key.Marshal())))
+			return nil, fmt.Errorf("key rejected for %q", c.User())
+		},
+	}
+	config.AddHostKey(hostPrivateKeySigner)
 
-    port := "2222"
-    if os.Getenv("PORT") != "" {
-        port = os.Getenv("PORT")
-    }
-    socket, err := net.Listen("tcp", ":"+port)
-    if err != nil {
-        panic(err)
-    }
+	port := "2222"
+	log.Println("Starting mock ssh server on " + port)
+	if os.Getenv("PORT") != "" {
+		port = os.Getenv("PORT")
+	}
+	socket, err := net.Listen("tcp", ":"+port)
+	if err != nil {
+		panic(err)
+	}
 
-    for {
-        conn, _ := socket.Accept()
-        go handle(conn, config)
-    }
+	for {
+		conn, _ := socket.Accept()
+		go handle(conn, config)
+	}
 }
